+++ internal/kogitoinfra_test.go
@@ -35,7 +35,7 @@ func TestFetchKogitoInfraInstance_InstanceFound(t *testing.T) {
 		},
 	}
 	cli := test.NewFakeClientBuilder().AddK8sObjects(kogitoInfra).Build()
-	context := &operator.Context{
+	context := operator.Context{
 		Client: cli,
 		Log:    test.TestLogger,
 		Scheme: meta.GetRegisteredSchema(),
@@ -50,7 +50,7 @@ func TestFetchKogitoInfraInstance_InstanceNotFound(t *testing.T) {
 	ns := t.Name()
 	name := "InfinispanInfra"
 	cli := test.NewFakeClientBuilder().Build()
-	context := &operator.Context{
+	context := operator.Context{
 		Client: cli,
 		Log:    test.TestLogger,
 		Scheme: meta.GetRegisteredSchema(),